move sload to individual mod
