+++ zkevm-circuits/src/evm_circuit/execution/sload.rs
