--- zkevm-circuits/src/evm_circuit/execution/storage.rs
@@ -1,3 +1,0 @@
-mod sload;
-
-pub(crate) use sload::SloadGadget;
